requirements.txt
Flask
python-dotenv
mysql-connector-python
pandas
scikit-learn
joblib
